5 - db-agent/pyproject.toml
[project]
name = "db-agent"
version = "0.1.0"
description = "A natural-language AI agent for managing SQL databases"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]
